# Unroll some inner loops