These unrolled loops improve performance significantly. Can include a benchmark if needed?