@donovanhide 2015-05-09
`benchcmp` output for latest commits
```benchmark                  old ns/op     new ns/op     delta
BenchmarkKeyGeneration     132027        66607         -49.55%
BenchmarkSigning           132649        69645         -47.50%
BenchmarkVerification      267982        204192        -23.80%
```