@DavidVorick 2015-07-25
At the very least you don't need constant time operations in the verification code.