@agl 2015-08-26
This was very hard to review, but I believe that it's correct, thanks!

I dropped the change to cmove because constant-time requires that the trace of instructions and data fetches be equal.

(GitHub mightn't think that it has been merged because of the rebase to drop that change.)