@donovanhide 2015-08-26
Thanks for merging! 
Maybe one day the Go compiler SSA work might be able to do all the register and instruction order optimisation automatically, but I needed fast as possible signing for a project, so it seemed worth the effort.
> This was very hard to review
I feel like that with some of djb's code :-)